GUI side of Load/Save

+++ Tank_Game/src/main/java/model/gameObjects/Bullet.java
@@ -111,7 +111,7 @@ public class Bullet extends GameObject {
     public void deserialize(String data) {
         String[] list = data.split(",");
         image = new Image(getClass().getResource("/Images/" + list[1]).toString());
-        position = new Point(Integer.parseInt(list[2]), Integer.parseInt(list[3]));
+        position = new Point((int)Double.parseDouble(list[2]), (int)Double.parseDouble(list[3]));
         direction = Integer.parseInt(list[4]);
         height = Integer.parseInt(list[5]);
         width = Integer.parseInt(list[6]);
